Make the domain name optional
- Implement a guess_domain function that will return the first custom
domain (or the first sandboxed one if sanbox is set to True).
- MailgunCredentialsError is now raised by the __request function.
- The domain name, if provided, is now checked against the list of
available domain. If it is invalid a MailgunDomainError is raised.
- A caching system has been implemented. If use_cache is set to True,
any call to __request will first check if the same request has already
been issued and return its results without making an actual API call,
thus preventing multiple calls to the GET /domains endpoint when
initializing the client.